feat/#19: localStorage에 현황 저장하는 코드 Layout에서 한 번에 처리
