+++ src/components/common/Header/Header.tsx
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 
 import icBack from '../../../assets/icons/back.svg';
 import icClose from '../../../assets/icons/close.svg';
+import useGoBackToHistory from '../../../hooks/ui/useGoBackToHistory';
 import { currentViewState, historyState } from '../../../states/atom';
 import theme from '../../../styles/theme';
 
@@ -14,23 +15,13 @@ interface HeaderProps {
 
 function Header(props: HeaderProps) {
   const { back, title, close } = props;
-  const [currentView, setCurrentView] = useRecoilState(currentViewState);
-  const history = useRecoilValue(historyState);
+  const goBackToHistory = useGoBackToHistory();
 
   const closePopup = () => {
+    // 초기화
     localStorage.setItem('currentView', 'size-option');
+    localStorage.setItem('currentTab', 'bottom');
     window.close();
-  };
-
-  const goBackToHistory = () => {
-    document.body.style.width = '38rem';
-    document.body.style.height = '37.5rem';
-    const container = document.getElementById('app-container') as HTMLElement;
-    if (container) {
-      container.style.width = '38rem';
-      container.style.height = '37.5rem';
-    }
-    history && setCurrentView(history);
   };
 
   return (
