+++ src/components/common/Layout.tsx
@@ -1,5 +1,8 @@
 import { PropsWithChildren, ReactNode } from 'react';
+import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
+
+import { currentViewState } from '../../states/atom';
 
 import Header from './Header/Header';
 
@@ -11,6 +14,10 @@ interface LayoutProps {
 }
 
 function Layout(props: PropsWithChildren<LayoutProps>) {
+  const currentView = useRecoilValue(currentViewState);
+
+  localStorage.setItem('currentView', currentView);
+
   const { children, back, title, close } = props;
   return (
     <Styled.Root>
