+++ src/pages/Popup/size-option/index.tsx
@@ -35,13 +35,7 @@ function SizeOption() {
     if (mySize) {
       /** TODO : 추천받은 사이즈가 없으면 nosize 있으면 size-recommend */
     }
-    if (mySize) {
-      setCurrentView('nosize');
-      localStorage.setItem('currentView', 'nosize');
-    } else {
-      setCurrentView('cannotload');
-      localStorage.setItem('currentView', 'cannotload');
-    }
+    mySize ? setCurrentView('nosize') : setCurrentView('cannotload');
   };
 
   return (
